Modify CFG in realEstimateCodeSize to use internal memory regions.
The CFG in realEstimateCodeSize is being allocated on the stack.
This means that CFG edges are also stack allocated and this
memory layout must be specify each time a CFG edge is built.
Instead of specifying stackAlloc every single time a CFG edge is
built, we specify the memory region during the initialization of
the CFG. This commit contributes to the following issue:

Issue: #6099

Signed-off-by: Erick Ochoa <eochoa@ualberta.ca>