Add tcp resources
Minimal changes to allow for a tcp functionality
to build. Build changes parallel those required
for jitaas, but do not include the ssl and other larger
changes that jitaas do compared to master.

TCP uses protobuf, protobufs currently not integrated into
jvm build, but script is present for compiling the proto
template.

Signed-off-by: knewbury <knewbury@ualberta.ca>