Add CogniCrypt mode tcp resources
Allow for a tcp functionality
to build. Build changes parallel those required
for jitaas, but do not include the ssl and other larger
changes that jitaas do. Still necessary to include, since
it is not the exact resources that jitaas uses, however
similar.

TCP uses protobuf, protobufs currently not integrated into
jvm build, but script is present for compiling the proto
template.

Signed-off-by: Kristen Newbury <knewbury@ualberta.ca>