Update tcp dir
Previously the tcp dir had been its own git repo
Therefore when it was added it was accidentally a
submodule, with non tracked content. This was not
intended, so it has been readded in order to provide
the content. Also updated the files for licenses,
that were not previously in there, again bc it was
initially developed separate from OpenJ9

Signed-off-by: knewbury <knewbury@ualberta.ca>