Setup Cogni Transformation
Register the optimization into warm strategy
and basic setup to build it. The opt performs
a transformation that adds a check above
some suspicious code location that currently will
correspond to an api call.

The callsite is id'd by 3 info pieces:
caller, callee and bci in caller. The opt looks
both in the compilation unit orig level and also
each method that may be inlined in the compilee already.

Signed-off-by: Kristen Newbury <knewbury@ualberta.ca>