Enable vm cogni seed read
If in cognimode rossa will now setup
a connection with cognicrypt
and read some seeds from it.
Also implemented persistent info
for client which is needed in both
rossa and then jit, and also those seeds.

Lastly added raw data reading in socket
for reading the seeds in the stream.

Signed-off-by: Kristen Newbury <knewbury@ualberta.ca>